Remove language module. (See note)
I've discovered that keeping up with this was a pain in the ass. I later
realized that in-browser translation is likely to be more robust and
accurate than this module ever will. The main en_us.php file will
contain all language now, for ease of editing. But now I will be able to
put more "dynamic" language in there.